# backend/spec/openapi.yaml
# Code generated by tools/generate_openapi.py. DO NOT EDIT.
openapi: 3.1.0
info:
  title: Олимпиадный маршрут
  version: 0.1.0
servers:
  - url: /api
paths:
  /v1/healthz/live:
    get:
      tags:
        - health
      summary: Liveness
      operationId: getLiveness
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /v1/healthz/ready:
    get:
      tags:
        - health
      summary: Readiness
      operationId: getReadiness
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/catalog:
    get:
      tags:
        - route
      summary: Catalog
      operationId: getCatalog
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Catalog"
  /v1/auth/max:
    post:
      tags:
        - route
      summary: Login Max
      operationId: loginMax
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /v1/auth/demo:
    post:
      tags:
        - route
      summary: Login Demo
      operationId: loginDemo
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
  /v1/me:
    get:
      tags:
        - route
      summary: Me
      operationId: getMe
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
      security:
        - HTTPBearer: []
    put:
      tags:
        - route
      summary: Profile
      operationId: saveProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Profile"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - HTTPBearer: []
    delete:
      tags:
        - route
      summary: Remove Profile
      operationId: deleteProfile
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
      security:
        - HTTPBearer: []
  /v1/track:
    get:
      tags:
        - route
      summary: Track
      operationId: getTrack
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackResponse"
      security:
        - HTTPBearer: []
  /v1/track/{olympiad_id}:
    put:
      tags:
        - route
      summary: Add
      operationId: addToTrack
      security:
        - HTTPBearer: []
      parameters:
        - name: olympiad_id
          in: path
          required: true
          schema:
            type: string
            title: Olympiad Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    patch:
      tags:
        - route
      summary: Change
      operationId: updateTrack
      security:
        - HTTPBearer: []
      parameters:
        - name: olympiad_id
          in: path
          required: true
          schema:
            type: string
            title: Olympiad Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackUpdate"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      tags:
        - route
      summary: Remove
      operationId: removeFromTrack
      security:
        - HTTPBearer: []
      parameters:
        - name: olympiad_id
          in: path
          required: true
          schema:
            type: string
            title: Olympiad Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackResponse"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /v1/notifications:
    get:
      tags:
        - route
      summary: Notifications
      operationId: getNotifications
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
      security:
        - HTTPBearer: []
  /v1/demo/events:
    post:
      tags:
        - route
      summary: Demo Event
      operationId: createDemoEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoEvent"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - HTTPBearer: []
  /v1/notifications/{notification_id}/actions:
    post:
      tags:
        - route
      summary: Action
      operationId: actOnNotification
      security:
        - HTTPBearer: []
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: string
            title: Notification Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationAction"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
components:
  schemas:
    Benefit:
      properties:
        program_id:
          type: string
          title: Program Id
        admission_year:
          type: integer
          title: Admission Year
          default: 2026
        kind:
          type: string
          enum:
            - bvi
            - "100"
            - unknown
          title: Kind
        result:
          type: string
          title: Result
        diploma_grades:
          items:
            type: integer
          type: array
          title: Diploma Grades
        confirmation:
          type: string
          title: Confirmation
        explanation:
          type: string
          title: Explanation
        source:
          $ref: "#/components/schemas/Source"
      type: object
      required:
        - program_id
        - kind
        - result
        - confirmation
        - explanation
        - source
      title: Benefit
    Catalog:
      properties:
        programs:
          items:
            $ref: "#/components/schemas/Program"
          type: array
          title: Programs
        olympiads:
          items:
            $ref: "#/components/schemas/Olympiad"
          type: array
          title: Olympiads
        snapshot_date:
          type: string
          title: Snapshot Date
          default: "2026-09-30"
        notice:
          type: string
          title: Notice
          default: Правила приёма 2026 года — ориентир. Условия вашего года нужно проверить заново.
        demo_enabled:
          type: boolean
          title: Demo Enabled
        bot_url:
          anyOf:
            - type: string
            - type: "null"
          title: Bot Url
      type: object
      required:
        - programs
        - olympiads
        - demo_enabled
      title: Catalog
    DemoEvent:
      properties:
        olympiad_id:
          type: string
          title: Olympiad Id
        kind:
          type: string
          enum:
            - registration
            - rule_change
          title: Kind
          default: registration
      type: object
      required:
        - olympiad_id
      title: DemoEvent
    Error:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Details
      type: object
      required:
        - code
        - message
      title: Error
    ErrorResponse:
      properties:
        error:
          $ref: "#/components/schemas/Error"
      type: object
      required:
        - error
      title: ErrorResponse
    Event:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        kind:
          type: string
          enum:
            - registration
            - stage
          title: Kind
        starts_at:
          anyOf:
            - type: string
            - type: "null"
          title: Starts At
        deadline:
          anyOf:
            - type: string
            - type: "null"
          title: Deadline
        source:
          $ref: "#/components/schemas/Source"
      type: object
      required:
        - id
        - title
        - kind
        - source
      title: Event
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Health:
      properties:
        status:
          type: string
          enum:
            - ok
          title: Status
          default: ok
      type: object
      title: Health
    HealthResponse:
      properties:
        data:
          $ref: "#/components/schemas/Health"
      type: object
      required:
        - data
      title: HealthResponse
    LoginRequest:
      properties:
        init_data:
          type: string
          maxLength: 16384
          minLength: 1
          title: Init Data
      type: object
      required:
        - init_data
      title: LoginRequest
    Me:
      properties:
        id:
          type: string
          title: Id
        is_demo:
          type: boolean
          title: Is Demo
        profile:
          $ref: "#/components/schemas/Profile"
      type: object
      required:
        - id
        - is_demo
        - profile
      title: Me
    Notification:
      properties:
        id:
          type: string
          title: Id
        olympiad_id:
          type: string
          title: Olympiad Id
        kind:
          type: string
          title: Kind
        state:
          type: string
          title: State
        due_at:
          type: number
          title: Due At
        deadline:
          type: number
          title: Deadline
        is_demo:
          type: boolean
          title: Is Demo
        sent_at:
          anyOf:
            - type: number
            - type: "null"
          title: Sent At
        title:
          type: string
          title: Title
        text:
          type: string
          title: Text
        error:
          anyOf:
            - type: string
            - type: "null"
          title: Error
      type: object
      required:
        - id
        - olympiad_id
        - kind
        - state
        - due_at
        - deadline
        - is_demo
        - sent_at
        - title
        - text
        - error
      title: Notification
    NotificationAction:
      properties:
        action:
          type: string
          enum:
            - registered
            - snooze
            - remove
            - read
          title: Action
      type: object
      required:
        - action
      title: NotificationAction
    NotificationList:
      properties:
        items:
          items:
            $ref: "#/components/schemas/Notification"
          type: array
          title: Items
      type: object
      required:
        - items
      title: NotificationList
    Ok:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
      type: object
      title: Ok
    Olympiad:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        profile:
          type: string
          title: Profile
        subject:
          type: string
          enum:
            - math
            - informatics
          title: Subject
        kind:
          type: string
          enum:
            - vsosh
            - listed
          title: Kind
        season:
          type: string
          title: Season
          default: 2026/27
        grades:
          items:
            type: integer
          type: array
          title: Grades
        description:
          type: string
          title: Description
        registration_url:
          type: string
          title: Registration Url
        source:
          $ref: "#/components/schemas/Source"
        benefits:
          items:
            $ref: "#/components/schemas/Benefit"
          type: array
          title: Benefits
        events:
          items:
            $ref: "#/components/schemas/Event"
          type: array
          title: Events
        level:
          type: string
          title: Level
          default: Уточняется для сезона 2026/27
      type: object
      required:
        - id
        - name
        - profile
        - subject
        - kind
        - grades
        - description
        - registration_url
        - source
        - benefits
        - events
      title: Olympiad
    Profile:
      properties:
        grade:
          type: integer
          maximum: 11.0
          minimum: 9.0
          title: Grade
          default: 10
        admission_year:
          type: integer
          maximum: 2035.0
          minimum: 2026.0
          title: Admission Year
          default: 2028
        subjects:
          items:
            type: string
            enum:
              - math
              - informatics
          type: array
          maxItems: 2
          minItems: 1
          title: Subjects
        program_ids:
          items:
            type: string
          type: array
          maxItems: 2
          title: Program Ids
        timezone:
          type: string
          title: Timezone
          default: Europe/Moscow
        notifications_enabled:
          type: boolean
          title: Notifications Enabled
          default: false
        quiet_start:
          type: integer
          maximum: 23.0
          minimum: 0.0
          title: Quiet Start
          default: 22
        quiet_end:
          type: integer
          maximum: 23.0
          minimum: 0.0
          title: Quiet End
          default: 8
        consent:
          type: boolean
          title: Consent
          default: false
      type: object
      title: Profile
    Program:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        short_name:
          type: string
          title: Short Name
        university:
          type: string
          title: University
          default: НИУ ВШЭ
        campus:
          type: string
          title: Campus
          default: Москва
        description:
          type: string
          title: Description
        url:
          type: string
          title: Url
          default: https://ba.hse.ru/
      type: object
      required:
        - id
        - name
        - short_name
        - description
      title: Program
    SessionResponse:
      properties:
        token:
          type: string
          title: Token
        expires_at:
          type: number
          title: Expires At
        user:
          $ref: "#/components/schemas/Me"
      type: object
      required:
        - token
        - expires_at
        - user
      title: SessionResponse
    Source:
      properties:
        url:
          type: string
          title: Url
        title:
          type: string
          title: Title
        checked_at:
          type: string
          title: Checked At
          default: "2026-09-30"
        note:
          type: string
          title: Note
          default: ""
      type: object
      required:
        - url
        - title
      title: Source
    TrackEntry:
      properties:
        olympiad_id:
          type: string
          title: Olympiad Id
        status:
          type: string
          enum:
            - planned
            - registered
            - completed
          title: Status
        added_at:
          type: number
          title: Added At
      type: object
      required:
        - olympiad_id
        - status
        - added_at
      title: TrackEntry
    TrackResponse:
      properties:
        items:
          items:
            $ref: "#/components/schemas/TrackEntry"
          type: array
          title: Items
      type: object
      required:
        - items
      title: TrackResponse
    TrackUpdate:
      properties:
        status:
          type: string
          enum:
            - planned
            - registered
            - completed
          title: Status
      type: object
      required:
        - status
      title: TrackUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
tags:
  - name: health
    description: Service health checks.